refactor: fix typo, and update barrel files
